@@ -13,22 +13,22 @@
     "private": true,
     "dependencies": {
         "@angular-material-extensions/select-country": "^7.0.1",
-        "@angular/animations": "~14.1.3",
-        "@angular/cdk": "^14.1.3",
-        "@angular/common": "~14.1.3",
-        "@angular/compiler": "~14.1.3",
-        "@angular/core": "~14.1.3",
-        "@angular/forms": "~14.1.3",
-        "@angular/material": "^14.1.3",
-        "@angular/material-moment-adapter": "^14.1.3",
-        "@angular/platform-browser": "~14.1.3",
-        "@angular/platform-browser-dynamic": "~14.1.3",
-        "@angular/router": "~14.1.3",
+        "@angular/animations": "~14.2.0",
+        "@angular/cdk": "^14.2.0",
+        "@angular/common": "~14.2.0",
+        "@angular/compiler": "~14.2.0",
+        "@angular/core": "~14.2.0",
+        "@angular/forms": "~14.2.0",
+        "@angular/material": "^14.2.0",
+        "@angular/material-moment-adapter": "^14.2.0",
+        "@angular/platform-browser": "~14.2.0",
+        "@angular/platform-browser-dynamic": "~14.2.0",
+        "@angular/router": "~14.2.0",
         "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
         "@fortawesome/fontawesome-free": "^6.1.2",
         "@ngx-translate/core": "^14.0.0",
         "@ngx-translate/http-loader": "^7.0.0",
-        "@schematics/angular": "^14.1.3",
+        "@schematics/angular": "^14.2.0",
         "@swimlane/ngx-charts": "^20.1.0",
         "@syncfusion/ej2": "^20.2.45",
         "jquery": "^3.6.0",
@@ -43,15 +43,15 @@
         "svg-country-flags": "^1.2.10"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^14.1.3",
-        "@angular-devkit/build-webpack": "^0.1401.3",
+        "@angular-devkit/build-angular": "^14.2.0",
+        "@angular-devkit/build-webpack": "^0.1402.0",
         "@angular-eslint/builder": "14.0.3",
         "@angular-eslint/eslint-plugin": "14.0.3",
         "@angular-eslint/eslint-plugin-template": "14.0.3",
         "@angular-eslint/schematics": "14.0.3",
         "@angular-eslint/template-parser": "14.0.3",
-        "@angular/cli": "~14.1.3",
-        "@angular/compiler-cli": "~14.1.3",
+        "@angular/cli": "~14.2.0",
+        "@angular/compiler-cli": "~14.2.0",
         "@bartholomej/ngx-translate-extract": "^8.0.2",
         "@types/d3-scale": "^4.0.2",
         "@types/d3-selection": "^3.0.3",
